+++ alacritty_terminal/src/event.rs
@@ -79,7 +79,7 @@ pub trait OnResize {
 }
 
 /// Event Loop for notifying the renderer about terminal events.
-pub trait EventListener {
+pub trait EventListener: Send + 'static {
     fn send_event(&self, _event: Event) {}
 }