+++ alacritty_terminal/src/tty/unix.rs
@@ -144,7 +144,7 @@ fn default_shell(pw: &Passwd<'_>) -> Program {
     let shell_name = pw.shell.rsplit('/').next().unwrap();
     let argv = vec![String::from("-c"), format!("exec -a -{} {}", shell_name, pw.shell)];
 
-    Program::WithArgs { program: "/bin/bash".to_owned(), args: argv }
+    Program::WithArgs { program: "/bin/bash".to_owned(), args: argv, input: None }
 }
 
 #[cfg(not(target_os = "macos"))]